Fix compile error on resistive screens
