+++ CapacitiveTouch.h
@@ -1,0 +1,11 @@
+#include <Wire.h>
+#include <FT6236.h>
+namespace FreeTouchDeck 
+{
+FT6236 ts = FT6236(); 
+void touch_calibrate()
+{
+    return;
+}
+
+}
