+++ ResistiveTouch.h
@@ -1,0 +1,77 @@
+namespace FreeTouchDeck
+{
+  /**
+* @brief This function presents the user with 4 points to touch and saves
+         that data to a claibration file.
+*
+* @param none
+*
+* @return none
+*
+* @note If USECAPTOUCH is defined we do not need to calibrate touch
+*/
+  void touch_calibrate()
+  {
+    uint16_t calData[5];
+    uint8_t calDataOK = 0;
+
+    // check if calibration file exists and size is correct
+    if (ftdfs->exists(CALIBRATION_FILE))
+    {
+      if (REPEAT_CAL)
+      {
+        // Delete if we want to re-calibrate
+        ftdfs->remove(CALIBRATION_FILE);
+      }
+      else
+      {
+        fs::File f = ftdfs->open(CALIBRATION_FILE, "r");
+        if (f)
+        {
+          if (f.readBytes((char *)calData, 14) == 14)
+            calDataOK = 1;
+          f.close();
+        }
+      }
+    }
+
+    if (calDataOK && !REPEAT_CAL)
+    {
+      // calibration data valid
+      tft.setTouch(calData);
+    }
+    else
+    {
+      // data not valid so recalibrate
+      tft.fillScreen(TFT_BLACK);
+      tft.setCursor(20, 0);
+      tft.setTextFont(2);
+      tft.setTextSize(1);
+      tft.setTextColor(TFT_WHITE, TFT_BLACK);
+
+      tft.println("Touch corners as indicated");
+
+      tft.setTextFont(1);
+      tft.println();
+
+      if (REPEAT_CAL)
+      {
+        tft.setTextColor(TFT_RED, TFT_BLACK);
+        tft.println("Set REPEAT_CAL to false to stop this running again!");
+      }
+
+      tft.calibrateTouch(calData, TFT_MAGENTA, TFT_BLACK, 15);
+
+      tft.setTextColor(TFT_GREEN, TFT_BLACK);
+      tft.println("Calibration complete!");
+
+      // store data
+      fs::File f = ftdfs->open(CALIBRATION_FILE, "w");
+      if (f)
+      {
+        f.write((const unsigned char *)calData, 14);
+        f.close();
+      }
+    }
+  }
+}
